Remove forward declaration of typedefs
Fixes #49 by removing the section which forward declares typedefs in the
`api.h.mako` template. First, the forward declarations are unnecessary
because the types are already defined where they appear in the header.
Second, when compiling with `clang` the following warning is emitted:

```
warning: redefinition of typedef 'ur_base_properties_t' is a C11 feature
```