ur: Remove user event creation
OpenCL user events, which `urEventCreate` was an analogue of, allow the
application to force a previously enqueued command to wait for the user
event to be signalled before execution can begin. User events are also
the only type of event that can synchronize across contexts, requiring a
separate special case code path to implement. Lastly, there is no
analogue to `clSetUserEventStatus` which actually enables the
application to change a user events status, making this feature
incomplete.

Until such a time as they are deemed absolutely necessary, with no
viable alternative proposed, this patch removes support for user events.

Fixes #15.