add function ids to the spec
XPTI function tracing requires "a stable ABI function id" that
is used to programmatically identify a traced entry point.
This patch adds a requirement for functions to have an unique id,
populates that field for existing functions and creates an enum
with each function id. It also adds a script that automatically
adds those ids if they are missing.