[ur] Fix linking against libstdc++fs
This patch replaces changes the logic used to specify linking against
`libstdc++fs.a` required to use early iterations of `std::filesystem`.
Previously CMake checked for the existence of the `<filesystem>` header
to enable linking the static library, however in GCC 8 this header
exists while still requiring manual linking against the static library.
Now we check if we are compiling with `libstdc++` and always manually
link against the static library. Fixes #451.