[CTS][CUDA] Ignore adapter warnings in conformance tests
Adapter warnings are supported in the UR spec for every
entrypoint. This commit updates the CTS to not fail
when a warning is returned. An environment variable was
added to optionally print the warnings to stderr.

A few fixes to the CUDA adapter were also added as part of this
commit:

- Always return a valid event for urEnqueueUSMPrefetch and
urEnqueueUSMAdvise.
- urAdapterGetLastError no longer ignores the error parameter.