[SPEC] Change PCI address query from string to struct
The `UR_DEVICE_INFO_PCI_ADDRESS` query previously returned a string of
the form `"DDDD:BB:SS.F"` whereas Level Zero and OpenCL (via the
`cl_khr_pci_bus_info` extension) return structs with each field
separated out. Additionally, CUDA and HIP both support querying the
domain, bus, and device ID's individually.

For applications which want to make decisions based on a devices PCI
address requiring to parse a string is not ideal. As such, this patch
changes the return type of the `UR_DEVICE_INFO_PCI_ADDRESS` to a struct
to more closely align with the Level Zero and OpenCL specifications.
Adapter implementations are also updated to align with the spec change.