[L0] fix caching internal events
When the L0 adapter enqueues an operation, it creates an event for
its own bookkeeping needs and to optionally notify the application
about completion. These events are periodically cleaned up and
the event objects themselves are cached to be reused. Or at
least they are supposed to be. If the enqueue does not specify
a user event (i.e., the event is L0 internal) the event
is destroyed and the L0 event handle is simply leaked in
urEventReleaseInternal. This means that every `MaxNumEventsPerPool`
enqueue calls (256 by default) a new event pool is created.

This patch fixes the above problem by making such internal events
own the native L0 handles, enabling them to be cached.