Optimization for SYCL out of order queue
For regular commandlist, we want to give the events device scope thus avoiding L3 flush.
To achieve this, I have a list of all events with device scope being kept track on within the queue. When we get to the last one, we want that commandlist's event to have host scope and also to wait for all other device scoped events.

Signed-off-by: Zhang, Winston <winston.zhang@intel.com>