Refactor guess local worksize funcs
The guessLocalWorkSize func for cuda adapter was erroneously giving a factor
of 2 even if it creates a global dim which is not the same as the user range
without using any range rounding kernels. This reverts that. Also refactor the
HIP version a little bit to remove an unnecessary check.