[HIP] Check for null arg to urKernelSetArgMemObj
The Properties parameter to `urKernelSetArgMemObj` is optional,
and so may be `nullptr`. However the HIP adapter assumes it is set
currently.

Updated this to check for nullptr, and if so assume that the buffer
is read/write.

Updated CTS match file so that the test executable doesn't segfault
early (CTS passes nullptr properties) and more tests run but fail.