{UR][L0] Return device version based on DeviceIpVersion
When being queried a device version:
- CUDA adapter returns the compute capability version of the device,
- HIP adapter returns the architecture name.

Both of these are what define a GPU architecture to compile binaries for
and can be used with GPU compilers, this is useful information for
applications and developers.

On L0/Intel side, prior to this change, it returns the L0 API version.
This information is already returned in BACKEND_RUNTIME_VERSION and
DRIVER_VERSION queries.

With this change, we align the adapters behavior with CUDA and HIP, and
return the Intel device architecture version that can be used with ocloc
compiler.