Terminate handles and flags regex strings with EOL character
This patch fixes undesired regex matching when the names of types
contain `handle_t` or `flags_t`, but do not end after the `_t`.

For example, an enum called `$x_exp_external_mem_handle_type_t`
shouldn't match to a handle type, as it is an enum.

Altering the regex strings for `RE_HANDLE`, `RE_IPC`, and `RE_FLAGS` to
terminate with the regex `$` end-of-line character fixes the undesired
matching behaviour.