[Loader] Refactor ONEAPI_DEVICE_SELECTOR, add tests
This patch resolves a number of issues with the extant implementation of
the `ONEAPI_DEVICE_SELECTOR`:

* Filtering only being applied to devices within a platform, resulting
  in an incorrect list of devices being returned in a multi-platform,
  multi-device context.
* Device indices being counted only inside a platform, not globally,
  resulting in multiple devices with 0 index in the presence of
  multiple platforms.
* Lack of testing for non-hardware dependent configurations, e.g. unable
  to emulate filtering against hypothetical sets of devices.
* Parsing the ONEAPI_DEVICE_SELECTOR string each time
  `urDeviceGetSelected()` was called.

This new implementation was written with testing of the individaul terms
in the ONEAPI_DEVICE_SELECTOR BNF grammer as a design requirement. It
decouples the enumeration of devices from the filtering logic. Both
parsing and device enumeration are moved much earlier, into loader
initialization, and performed only once.