Use mutexes for cached events storing and loading
Make the retrieval and pushing of events onto the stack thread-safe
by locking it with mutexes.
Use these mutexes inside the queue, which is responsible for caching
the events.

Apply further minor changes, such as returning the result status
for event releases directly, usage of the proper UR assertions
and proper styling for variable names.