Associate queue with device in context
Making a native queue doesn't require hDevice to be non null, but this
associates the queue with a null device, even if hContext contains valid
devices.

This fixes the behaviour for a single device context, making the queue
be associated with the only device in the context. But there is a
question mark over whether this entry point makes sense for a multi
device UR context. Should we be able to construct a queue with a native
queue without knowing what the UR device associated with the queue is?
Probably not.