[CI] Add "loader" support to conformance testing
This expands our CI to test the loader; the dispatcher that is
used when multiple adapters are availabe.

The old "run on hardware" jobs should behave the same (they have
the loader tests disabled), but there is a new
`combined-opencl-level-zero` job that tests a build with both
OpenCL and Level Zero available.