[UR][Bindless] Initial implementation of bindless images for HIP
This PR implements the entry points in the HIP adapter for the Bindless Images extension.

Some features, while implemented, only partially pass the End-to-end tests for them in
DPC++ (the PR that enables the support and testing in DPC++: intel/llvm#16439).
This is because the HIP runtime doesn't support, for example, some parameter combinations
with some image memory types. The not-fully supported features are marked with TODOs in
the device info queries for further work.

Signed-off-by: Georgi Mirazchiyski <georgi.mirazchiyski@codeplay.com>

Co-authored-by: Peter Žužek <peter@codeplay.com>