Wrap linker flags on Windows for IntelLLLVM
The Intel C++ compiler requires linker flags to be wrapped, because
CMake passes them through the compiler driver.
Prefix linker options with `LINKER:`.
CMake will transorm it to the appropriate flag for the compiler driver:
(Nothing for MSVC or clang-cl, and /Qoption,link for icx), so this will
work for other compilers and for earlier CMake versions too.

Signed-off-by: Gergely Meszaros <gergely.meszaros@intel.com>