[CL] Support using old ICD loaders
While we recommend using the latest OpenCL-ICD-Loader version this is
not always possible, perhaps due to a system library being outdated with
no ability to update due to lacking admin privileges. This can cause the
OpenCL adapter library to silently fail to load or worse cause a crash.

This patch changes entry points adding in OpenCL 2.2 or later to be
loaded from the OpenCL-ICD-Loader at runtime instead of during startup
by the operating system. The entry point function pointers are stored in
the OpenCL adapter's `ur_adapter_handle_t` and all previous uses are
replaced with checked calls to these function pointers.

Previously `clGetExtensionFunctionAddressForPlatform` was being used,
however this is not guaranteed to return valid function pointers for
core OpenCL entry points even if they are present.