No longer expect segfaults in match files
A number of match files check for `Segmentation Fault`, which happens
when the test runner crashes. This change removes that check and
replaces it with checks for the tests that actually cause the segfault.
As the test runner no longer segfaults, more tests are executed and so
more non-segfaulting fails needed to be added to the match files.

In addition, a number of match files had the check for `Segmentation
Fault` removed if they didn't actually segfault.

Note that this will have no effect for the `ctest_parser.py` script, as
it ignores the match file logic.