Add try catches to anything that can throw
UR_CHECK_ERROR must be called within a try catch block, as it can throw.

For functions that don't return a `ur_result_handle_t`, and which might
throw, make sure these funcs are only called within try catches in the
caller funcs, and that they are not exposed to the user.