Add CTS macro for marking tests as known failures in code
The new `UUR_KNOWN_FAILURE_ON` macro can be used to skip tests on
devices where the test is known to fail. This can be done in the
following situations.

For all devices in an adapter:

```cpp
UUR_KNOWN_FAILURE_ON(uur::LevelZero{});
```

By substring match of the device name within and adapter:

```cpp
UUR_KNOWN_FAILURE_ON(uur::OpenCL{"Intel(R) UHD Graphics 770"});
```

> Note: This invocation is used in a few places in this patch to
> facilitate clean runs when using this decice.

In certain test suits, where there is no access to a device, the
platform name is used instead:

```cpp
UUR_KNOWN_FAILURE_ON(uur::CUDA{"NVIDIA CUDA BACKEND"});
```

When neither device or platform is available in a test suite, the name
is ignored and only the adapter backend is used to determine if the test
is a known failure.

The macro is variadic making it possible to specify known failures for
multiple adapters in a single place and multiple names can also be provided
per adapter:

```cpp
UUR_KNOWN_FAILURE_ON(
    uur::OpenCL{
        "Intel(R) UHD Graphics 750",
        "Intel(R) UHD Graphics 770",
    },
    uur::HIP{"Radeon RX 7700"},
    uur::NativeCPU{});
```