[L0] Remove zero size check from getImageRegionHelper
A SYCL bindless image test, that has ext_oneapi_copy call with copy
region width = 0, fails in UR level-zero adapter due to the check.

On the other hand, the test can run successfully on cuda adapter.
In order to make test behavior consistent across different adapters,
this PR leaves the check to the level-zero backend and the backend
is supposed to return success for this case.