Improve command-buffer update error code wording
The `UR_DEVICE_INFO_COMMAND_BUFFER_UPDATE_SUPPORT_EXP` query no longer
exists, so error defined for trying to make an updatable command-buffer
when a device doesn't support it needs reworded to say if no
capabilities are supported.

Additionally, for each individual update capability, specify that
an error is thrown if update is attempted using that characteristic.

UR CTS added to verify this.