Enabled the use of Copy engine with in order command lists
Added a createSyncPointBetweenCopyAndCompute() function which appends a
signal event to the previous command list to wait on. This can happen in
a scenario when we use Compute -> Copy -> Compute, when the last Compute
needs to wait for Copy to preserve in-order behaviour. For the
immediateAppend path, this works only if the counter-based events are
explicitly not used.