In-order path for OpenCL command-buffers
After the [spec bump of cl_khr_command_buffer to
0.9.7](https://github.com/KhronosGroup/OpenCL-Docs/),
in the OpenCL adapter we no longer need to worry about the in-order/out-of-order property
of the internal queue used on command-command-buffer creation matching
the queue used to enqueue the command-buffer.

We can therefore take advantage of the in-order flag passed on
UR command-buffer creation to use an in-order queue for command-buffer
creation, and omit using sync points.