Adjust backtrace addresses to call instruction (#17404)
`backtrace()` would return buffers with return addresses, which is one
instruction past the call. If that instruction belongs to other source
lines, then the llvm-symbolizer would give out wrong result.
Test will be added internally. It is hard to write test for this using
SYCL as SYCL does not generate code with that pattern.