Fixes for using `ONEAPI_DEVICE_SELECTOR` with linear ids (#18029)
Linear IDs (the `0` in `opencl:0`) did not function as intended when
there was multiple platforms; each platform started counting at zero.
This patch changes it so that they start counting at an appropriate
number that should match `urinfo`.

`urinfo` itself has also been updated to not display linear IDs when
the device list is being filtered, as they will not be correct. This
matches the behavior of sycl-ls and is different to the
`--no-linear-ids` which changes the format of the output.