Serialize submissions of the same command-buffer (#18295)
Update the UR command-buffer enqueue semantics such that command-buffer
submissions are ordered with a dependency on previous submissions of the
same command-buffer.

That is the same semantic as we currently specify for executable graph
objects in the SYCL-Graph extension. Allowing for a more efficient
SYCL-RT implementation on top of UR where less UR events need to be
created (future work).

OpenCL is the only adapter that needs updated to provide these
semantics, where we now track the `cl_event` returned from the most
recent submission of a command-buffer, so that it can be used as a
dependency of any future command-buffer submissions.